feat: exaspim version

+++ view/widgets/miscellaneous_widgets/q_scrollable_float_slider.py
@@ -14,7 +14,7 @@ class QScrollableFloatSlider(QSlider):
         return float(super().value()) / self.divisor
 
     def setMinimum(self, value):
-        return super().setMinimum(value * self.divisor)
+        return super().setMinimum(int(value * self.divisor))
 
     def setMaximum(self, value):
         return super().setMaximum(int(value * self.divisor))
